divided JS code into separate files which renders in index.js
