+++ static/js/carImag.js
@@ -1,0 +1,46 @@
+// document.addEventListener('DOMContentLoaded', function() {
+
+    // document.getElementById('user_cars').addEventListener('change', function() {
+    //     let selectedIndex = this.selectedIndex;
+
+    //     // Получаем выбранный элемент <option> по индексу
+    //     let selectedOption = this.options[selectedIndex];
+
+    //     // Получаем id выбранного элемента <option>
+    //     let selectedBrand = selectedOption.id;
+    //     let carImage = document.getElementById('current_car_img');
+    //     let loadingMsg = document.getElementById('loadingMsg');
+
+    //     loadingMsg.style.display = 'block';
+    //     carImage.style.display = 'none';
+
+    //     fetch('/static/img/default_brand/' + selectedBrand + '.jpg')
+    //         .then(function(response) {
+    //             if (response.ok) {
+    //                 return response.blob();
+    //             } else {
+    //                 throw new Error('Image not found ' + response.status);
+    //             }
+    //         })
+    //         .then(function(imageBlob) {
+    //             let imageUrl = URL.createObjectURL(imageBlob);
+    //             carImage.src = imageUrl;
+
+    //             loadingMsg.style.display = 'none';
+    //             carImage.style.display = 'block';
+    //         })
+    //         .catch(function(error) {
+    //             console.error('Error loading image:', error);
+    //             loadingMsg.textContent = 'Error loading image'
+    //         })
+    // })
+
+
+    // console.log('testing js connection');
+
+    
+
+    
+    
+// });
+
